Modularized component logic.
