+++ app/components/GettingStarted.jsx
@@ -1,0 +1,8 @@
+import React, { useContext } from 'react';
+import AddService from './AddService.jsx';
+
+const SignUp = () => {
+  return <AddService />;
+};
+
+export default SignUp;
